TASK: Refactor rootline handling
Instead of iterating over the rootline, the rootline is set up so that every
child knows the parent. If the child itself has no size configuration, the
calculated multiplier is passed on to the parent for calculation. In this way,
the complete calculation is stored out of the rootline class.

New classes had to be introduced for the columns of the containers and the
columns of the backend layout. A corresponding element is then added to the
rootline for these columns.

Related: #26